requirement.txt
pytorch==1.1.0
torchvision==0.3.0
numpy
scipy
matplotlib
scikit-image
graphviz

## python3.7

